ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.7.4
python-dotenv==1.0.1
asyncio-extras==1.3.2
